+++ docs/releasenotes.rst
@@ -20,6 +20,7 @@ V1.3.30:
 
 - fix: check curl availability in hec_wrapper.sh
 - fix: redirect to null any curl stderr/stdout in hec_wrapper.sh
+- fix: missing log_date function in hec_wrapper.sh
 
 ========
 V1.3.29: